Reworking the Material Button
The material button offer a handful of variations. These variations obey
a set of rules. For example, type selectors are mutually exclusive
["mdl-button--raised"|""|"mdl-button--icon"|"mdl-button--fab"] while
mdl-button--mini-fab depends on mdl-button--fab to work, mdl-button--fab
can only be accentuated, button has to be upgraded before used and so
on. It is important to encapsulate these rules to make the button
interface intuitive. One way to do it is to build buttons through a
factory. Variations for the same aspect are exposed using Enums. First,
the user select the builder and then for each aspect chooses the
suitable option from the closed set of alternatives.