Fixed Menu design issue.
After deeper reflection, it might not be a good idea to tightly couple
the action button and the menu together. Many problems will arise from
this design choice. Menus can only be activated by a restricted set of
buttons. It is better to leave the activation component outside the
scope of this component.
This implementation deviates too much from mdl and it would make updates
more complicated if the JS component is enhanced in a future revisions.