Fixed MenuItem not upgraded
componentHandler.upgradeElement(element) only operates on the very
element it is applied on. It does not operate on inner elements within.
Intuitively, the upgrade should be performed on the DOM tree that has
element as root but this is not the case. This issue is fixed by
upgrading each inner element individually.