Added js script to handle component injection.
Some parts of the layout like banners and navigation bars will be
included in virtually in demo page with slight variations. In addition
to make the page more bloated, it will induce a maintenance nightmare.

One straightforward solution is to provide a script that injects into
the page reusable parts into their appropriate location.

This script would make it possible to decompose the layout into reusable
chunks. This will help keep the code as dry as possible by avoiding
redundancy.