One module to rule them all
Making a module per demo page doesn't work very well performance wise. As the number of pages increases, the time needed to compile all the modules required for those pages grows linearly. This results in an increased compilation time for development as well as at packaging. The chosen solution to overcome this problem is to aggregate all Entrypoints inside a composite Entrypoint. The composite's job is to determine which actual Entrypoint to delagate the processing to. The composite stores the collection of actual Entrypoints. Each Entrypoint is mapped to a target url. When a match is found between the Entrypoint's target url and the current page's url, the composite delegates the handling of the current page to the selected Entrypoint. This implies that the demo application will have a single module and therefore a single compilation session and each page is handled by the appropriate Entrypoint based on a url mapping. That mapping should be updated whenever the location of the page is updated.