Adhere to coding standards by committing to sun's checkstyle
The library module code was reviewed and fixed to make sure it validates against sun checkstyle rules. Considering some rules can't be efficiently managed by Eclipse IDE and its formatting module, those rules where relaxed to keep the project manageable using the IDE. For example, when linking to javadoc classes, Eclipse adds automatically an import that is considered a bad practice by the sun rules. This rules is irrelevant in our case since the code will ultimately be compiled to javascript. Javadoc line return is a formatting limitation imposed by Eclipse, so it had to be relaxed in the style validation. And finally, Line length was expanded to 150 characters (eclipse recommandation).