Fixed the radio graphic not triggering the onValueChange event
In the original radio component, the native input element along with its label are the only actors in the radio component. Event propagation needs to take into account only these components. Howerver, in the material implementation the input element is not visible and does not receive the mouse input which becomes the responsibility of the radio button rings. Therefore, events should be positioned on the root level container.