Added demos for releases along side the snapshot demo
In order to make the releases demos live side by side, each demo is located inside a dedicated sub directory.
The snapshot demo is updated for each commit and is now living inside the snapshot dedicated directory.
For each push on dev, the snapshot directory is updated to reflect the changes applied.