Fixed issue, the checkbox does not react to browser events
The original GWT checkbox sinks events on the input element. This means that browser event handlers are attached to the input element. When an event handler is attached to the checkbox, the event is in fact attached to the input node. For example if the checkbox is set to listen for mouse click events, it is the input element that should be clicked to  actually trigger the handler. Since the input element is invisible, events handlers attached on the checkbox are never triggered. To fix this issue, the sinking of events should be applied on the checkbox container rather than the input node.