Making menu wrappers the source of item click events
MenuCombo and Dropdown both contain a Menu instance and it is that very component that manages item click events.
The source for the dispatched events are obviously that inner instance rather than the Dropdown or the Menucombo itself.
This implementation detail is counter intuitive and the end developper would never expect the source of the event to be something else rather than the MenuCombo instance itself.
The solution is to designate the MenuCombo or Dropdown as the source of the event.
When an item click event is set to handle more than one Dropdown, the source of events is simply compared against each Dropdown reference.