Making the Slider expose all the features provided by the SliderInput
The Slider component is made of a container (Slider) wrapped around an input range (SliderInput).
The input needs the container so it can be resized properly. Each time a service is needed from the input,
a reference to the input should be acquired. This is cumbersome and sometimes confusing especially when dealing with events.
The proposed solution is to completely shield the input from the outside and fully support its features by the container.